pyproject.toml
[project]
name = "ai-sales-assistant"
version = "0.1.0"
description = "AI Sales Assistant для малого бизнеса - автономный Telegram-бот для квалификации лидов"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.2",
    "aiogram>=3.22.0",
    "anthropic>=0.40.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tortoise-orm>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Игнорируем библиотеки без типов
[[tool.mypy.overrides]]
module = [
    "aiogram.*",
    "tortoise.*",
    "aerich.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # annotations
    "B",   # bugbear
    "A",   # builtins
    "COM", # commas
    "C4",  # comprehensions
    "DTZ", # datetime
    "ISC", # implicit-str-concat
    "ICN", # import-conventions
    "PIE", # pie
    "T20", # print
    "PT",  # pytest-style
    "Q",   # quotes
    "RET", # return
    "SIM", # simplify
    "ARG", # unused-arguments
    "PTH", # use-pathlib
    "ERA", # eradicate
    "PL",  # pylint
    "TRY", # tryceratops
    "RUF", # ruff-specific
]
ignore = [
    "ANN101",   # missing-type-self
    "ANN102",   # missing-type-cls
    "ANN401",   # dynamically-typed-expression (Any is needed for middleware)
    "COM812",   # trailing-comma-missing (conflicts with formatter)
    "ISC001",   # single-line-implicit-string-concatenation (conflicts with formatter)
    "RUF001",   # ambiguous-unicode-character-string (Russian project)
    "RUF002",   # ambiguous-unicode-character-docstring (Russian project)
    "RUF003",   # ambiguous-unicode-character-comment (Russian project)
    "RUF012",   # mutable-class-default (Tortoise ORM Meta classes)
    "PLR2004",  # magic-value-comparison (acceptable for simple cases)
    "TRY003",   # raise-vanilla-args (acceptable for simple cases)
    "TRY004",   # type-check-without-type-error (ValueError is ok)
    "TRY301",   # raise-within-try (acceptable for validation)
    "A005",     # builtin-module-shadowing (src.types is ok)
    "DTZ003",   # tz-naive-datetime (utcnow is fine for our use case)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__ are ok

